# go-image-server/docs/swagger.yaml
info:
  contact: {}
paths:
  /images:
    get:
      description: Get all image filenames from the static directory.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Unable to read directory
          schema:
            type: string
      summary: List all images
      tags:
      - images
  /images/{filename}:
    delete:
      description: Deletes an image file from the static directory by its filename.
      parameters:
      - description: Image file name
        in: path
        name: filename
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: File name required
          schema:
            type: string
        "404":
          description: Image not found
          schema:
            type: string
        "500":
          description: Failed to delete file
          schema:
            type: string
      summary: Delete an image
      tags:
      - images
    get:
      description: Serve image by filename, with optional resizing (w, h, fit, auto=format)
      parameters:
      - description: Image file name
        in: path
        name: filename
        required: true
        type: string
      - description: Width
        in: query
        name: w
        type: integer
      - description: Height
        in: query
        name: h
        type: integer
      - description: Fit mode (crop)
        in: query
        name: fit
        type: string
      - description: Auto format (format)
        in: query
        name: auto
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: File name required
          schema:
            type: string
        "404":
          description: Image not found
          schema:
            type: string
      summary: Get image file
      tags:
      - images
  /restart:
    get:
      description: Stops and restarts the feh slideshow viewer.
      produces:
      - text/plain
      responses:
        "200":
          description: feh restarted.
          schema:
            type: string
      summary: Restart the feh slideshow
      tags:
      - feh
  /start:
    get:
      description: Launches the feh slideshow viewer on the configured uploads directory.
      produces:
      - text/plain
      responses:
        "200":
          description: feh started.
          schema:
            type: string
      summary: Start the feh slideshow
      tags:
      - feh
  /stop:
    get:
      description: Stops the running feh process if active.
      produces:
      - text/plain
      responses:
        "200":
          description: feh stopped.
          schema:
            type: string
      summary: Stop the feh slideshow
      tags:
      - feh
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file via multipart form data.
      parameters:
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: File uploaded successfully
          schema:
            type: string
        "400":
          description: Missing file
          schema:
            type: string
        "405":
          description: Invalid request method
          schema:
            type: string
      summary: Upload image
      tags:
      - upload
swagger: "2.0"